requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
Django
django-apscheduler==0.6.2
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzlocal==5.0.1
